chore: Show that promise deadlock doesn't repro.
This is a legacy test case that ~a few years ago would promise lock, but
that a populate refactoring fixed.
